feature/cocktail-sort: Major Fixed
Previous modules are now bundled inside a large main.js file inside modules folder.

Initialization of the modules have been converted into conditionals that depend on the page name.
Page name variable has been added to some files. They are then retrieved from a meta tag inside the head.

Suggested changes are addressed.

Working initializations from the main.js:
- Likes
- Sort and category
- Sticky
- Random
- Profile
- Comments
- Cocktail edit
- Search

Admin functions are set on pause and are currently commented out as Brilliant Beaver has made some changes.

Known issue:
Exhaustion of events on home page. All functions related to homepage work, however if there is and overload of (specifically) sorting and category, the initialization may cause looping.

This shall be a simple fix for later itteration.
